requirements.txt
langchain_core==0.3.7
mlx_lm==0.18.2
mlx_whisper==0.3.0
python-dotenv==1.0.1
python-telegram-bot==21.6
yt_dlp==2024.9.27
